requirements.txt
# Flask and related extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.2

# Database driver for PostgreSQL
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Password hashing
Werkzeug==2.3.7

# Logging and debugging
loguru==0.7.2

# Report generation
matplotlib==3.8.0
reportlab==4.0.4

# CSV handling
pandas==2.1.1

# Time zone handling (for Python 3.9+)
pytz==2023.3

# Deployment utilities
gunicorn==20.1.0

# Testing and development tools
pytest==7.4.2
flake8==6.1.0
